add, update, delete by user added

+++ 02_The_Paractical_Application_Server/controllers/feed.js
@@ -4,9 +4,10 @@ const path = require('path');
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
-  const currentPage = req.query.page;
+  const currentPage = req.query.page || 1;
   const perPage = 2;
   let totalItems;
   Post.find()
@@ -17,9 +18,10 @@ exports.getPosts = (req, res, next) => {
         .skip((currentPage - 1) * perPage)
         .limit(perPage);
     })
-    .then((records) => {
+    .then((posts) => {
       res.status(200).json({
-        posts: records,
+        message: 'Fetched posts successfully.',
+        posts: posts,
         totalItems: totalItems,
       });
     })
@@ -34,37 +36,40 @@ exports.getPosts = (req, res, next) => {
 exports.createPost = (req, res, next) => {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
-    const error = new Error('Validation Failed !');
+    const error = new Error('Validation failed, entered data is incorrect.');
     error.statusCode = 422;
     throw error;
   }
-
   if (!req.file) {
-    const error = new Error('No image provided !');
+    const error = new Error('No image provided.');
     error.statusCode = 422;
     throw error;
   }
-
   const imageUrl = req.file.path;
   const title = req.body.title;
   const content = req.body.content;
-
+  let creator;
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: {
-      name: 'Danish Riaz',
-    },
+    creator: req.userId,
   });
-
   post
     .save()
-    .then((record) => {
-      console.log(record);
+    .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then((result) => {
       res.status(201).json({
         message: 'Post created successfully!',
-        post: record,
+        post: post,
+        creator: { _id: creator._id, name: creator.name },
       });
     })
     .catch((err) => {
@@ -75,18 +80,16 @@ exports.createPost = (req, res, next) => {
     });
 };
 
-exports.getPostById = (req, res, next) => {
+exports.getPost = (req, res, next) => {
   const postId = req.params.postId;
   Post.findById(postId)
-    .then((record) => {
-      if (!record) {
-        const err = new Error('No Post Found.');
-        err.statusCode = 404;
-        throw err;
+    .then((post) => {
+      if (!post) {
+        const error = new Error('Could not find post.');
+        error.statusCode = 404;
+        throw error;
       }
-      res.status(200).json({
-        post: record,
-      });
+      res.status(200).json({ message: 'Post fetched.', post: post });
     })
     .catch((err) => {
       if (!err.statusCode) {
@@ -98,7 +101,6 @@ exports.getPostById = (req, res, next) => {
 
 exports.updatePost = (req, res, next) => {
   const postId = req.params.postId;
-  console.log(postId);
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
     const error = new Error('Validation failed, entered data is incorrect.');
@@ -121,6 +123,11 @@ exports.updatePost = (req, res, next) => {
       if (!post) {
         const error = new Error('Could not find post.');
         error.statusCode = 404;
+        throw error;
+      }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
         throw error;
       }
       if (imageUrl !== post.imageUrl) {
@@ -151,12 +158,23 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
       // Check logged in user
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then((result) => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then((result) => {
       res.status(200).json({ message: 'Deleted post.' });
     })
     .catch((err) => {
